feat: 이메일 인증 기능 간결화
